Now when the generate button is hit the instnace that appear in Vue.js from the data are only after the start date that was inputed
In order for this to be created, a Generate Report button, a logs variable in Vue.js, and a saveLogs method was created. When the Generate report button is cicked the saveLogs method is started. In the saved logs method a variable called link is created. This variable is set to the standard link for all transplanting log. Then if startDate is not empty then the link variable is added to inorder to create a link that will get all transplanting logs created after that start date. Then that link is used in the axios call. In order to do this, the methodstimestampToYMD and YMDToTimestamp were created. This methods change YMD time to timestamp time and vice versa.