Adds user/crop/area mapping functions (#211)
* added APIRequestFunction and first test to resources folder

* took out api function that is not complete yet

* added functions getCropMap, getFieldMap, getUserMap, and getMap.
The first three functions call getMap. Then get map uses there url and id to create a map that has the id to the name of the crop,field, user, or other piece of information that would be useful in a map structure

* added function for getting a map that maps crop, user, or field to an id. Added test for these functions in cypress. Also added tests for the functions that mapped id to crop,user, or field.

* fixed discription of length test for getCropToIDMap

* all test for crop,field, or user to id and vice versa work correctly

* put in requested changes.

This included changing the crop url and having each function only have one test(combining the old tests). Also put all the tests for the map functions into one context.

* Removed incorrectly merged test code, renamed orig db to work with setDB.bash script

Co-authored-by: josieecook <cookjo@dickinson.edu>
Co-authored-by: IrisSC <shakerci@dickinson.edu>