Typo fixes and other suggestions for FarmData school activities 03-07.
This commit is intended to be part of a _draft_ pull request.  It is
not intended to be merged.  The docx files are all in Track Changes
mode, and some have comments added.  The intention is that a
maintainer will examine the changes and approve them with alterations
as needed.  After that, a (non-draft) pull request will be created in
which the files are not in Track Changes mode and with all of the
corresponding pdf files updated with the relevant changes.