Typo fixes and other suggestions for FarmData school activities 03-07. (#385)
* Typo fixes and other suggestions for FarmData school activities 03-07.

This commit is intended to be part of a _draft_ pull request.  It is
not intended to be merged.  The docx files are all in Track Changes
mode, and some have comments added.  The intention is that a
maintainer will examine the changes and approve them with alterations
as needed.  After that, a (non-draft) pull request will be created in
which the files are not in Track Changes mode and with all of the
corresponding pdf files updated with the relevant changes.

* Finalizing changes to FarmData2 school Activities 03-07

This commit finalizes previously-submitted changes to FarmData2 school
Activities 03-07. See commit b3b7a08e3fcf5a4212cf0c6b385dfac2db68c611
for details.

In this commit, the Accept All Changes and Stop Tracking functionality
and Microsoft Word was used to finalize the docx documents. All
comments were deleted. The maintainer's suggestions were
incorporated. New PDF files were generated from the docx files.