add quotation marks to html attribute values
These quotation marks are not needed for correctness,
but it is better to include them for consistency and
for other reasons.
See e.g.
https://stackoverflow.com/questions/6495310/do-you-quote-html5-attributes
It would be even better would to keep a consistent style with either
all single quotes or all double quotes,
but in this commit we are just adding quotation marks where
they are missing altogether.