Refactoring Pre-existing Seeding Report e2e Test (#514)
* Cypress test changes for editing/deleting buttons
Fixing the test, but it currently is not accepting a valid column and row

* Additions to Seeding Report and Refactor Cypress Tests
Seeding Report had some data-cy tags added to the summary tables to easily access that data for testing purposes. All Cypress tests that were malfunctioning due to CustomTable changes are now working. All Cypress e2e tests now work. Some additional ones need to be added to account for API calls.

* Cypress Tests: Testing API and Filter Selection
Added tests to make sure that the filter dropdowns are being populated by the API calls to the crop map and area map.
Added tests to make sure selections in one filter dropdown impact the options in the other filter dropdowns.
Added coverge for visible columns when the edit button makes the save and cancel headers visible.

* Refactored Cypress Test File Structure
Removed a good amount of the logins from the beforeEach calls made inside of contexts and allowed that to always occur in the describe before each test. This seems to have helped significantly reduce the amount of times a potential 403 error might occur at the cost of increased test time since the beforeEach in the describe runs for each test.
Additionally, some of the API calls are currently incomplete because I do not know where exactly to make those API calls to.

* Fixed API Issue
The first test was causing the rest of a context to fail since it was reloading the page.
Removed commented code.
Made sure everything was properly indented and spaced.

* Removed some redundant tests
1. Removed a repeated tests that tested filter behavior

* Made Requested Changes to the Seeding Report e2e Test
1. Added two more data-cys to the Seeding Report html page in order to verify if the text denoting no longs are there for that seeding type is present or not
2. Removed a context and combined some of its contents to another context because they overlap.
3. Changed the edit/deleting tests at the end of the e2e file.
4. (Hopefully) Fixed all spacing and indentation concerns