Reorganization of the farmdata2 modules (#602)
__Pull Request Description__

This PR reorganizes the code for the FarmData2 modules and api. There is
now a `farmdata2` directory that contains all of the api and application
code for FarmData2 in a single location. In that directory there are
directories for `farmdata2_api` and `farmdata2_modules`.

In the `farmdata2_modules` directory there is a directory for each of
the modules (e.g. `fd2_barn_kit`, `fd2_example`, etc) and the
`resources` directory.

In the directory for each module there are the `.info` and `.module`
files along with any other files necessary to create the Drupal module.
There is then a directory for each sub-tab that appears in the farmOS
UI. For example, in the `fd2_example` folder there are folders for
`Info`, `Vars`, `Maps`, `API`, `Cache` and `UI`.

In each of the sub-tab folders is the `.html` file that contains the
code for the sub-tab along with a collection of `.spec.js` files that
hold the e2e tests for the sub-tab.

This organization should make it easier to create new sub-tabs and to
organize the tests for each sub-tab.

The testing and documentation scripts were moved up to the `farmdata2`
directory so that they can now be used both with the modules and the
api.

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.