Updated fd2-down.bash to remove dev container. (#603)
__Pull Request Description__

Updates fd2-down.bash to remove the dev container when FarmData2 is shut
down. There was an issue when running in WSL that is addressed by this
fix. The fix is a little unsatisfactory in that the writeable later in
the dev container is lost. Thus, any local installs done are lost when
FarmData2 is taken down. If this presents a challenge for lots of
developers then this issue should be revisited.

This closes #584.

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.