Added id to PayLoad of CustomTable edit-clicked event (#604)
__Pull Request Description__

This PR adds the id of the row being edited in the CustomTableComponent
to the `edit-clicked` event. It makes corresponding changes in the unit
tests for the component and in the UI example page and its tests.

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.

Co-authored-by: Chris <31524934+FutzMonitor@users.noreply.github.com>