Added ENV var to prevent WSL warning for Codium (#611)
__Pull Request Description__

When running VSCodium on Windows/WSL it detected that it was running in
Docker and suggested using it under WSL directly. The added environment
variable silences this warning.

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.