Reorganizes the Documentation across the repo. (#614)
__Pull Request Description__

This PR reorganizes the documentation as follows:
- CONTRIBUTING focuses on making a first contribution.
- A RESOURCES page is added serves as a pointer to all other
documentation.
- READMEs for each type of work (modules, libraries, api, etc.) are kept
with their code and linked from RESOURCES.
- A root level `docs` folder is added to hold "big picture" docs that
are not tied to a specific directory and are linked from RESOURCES.

- This is not quite complete but is largely complete. After this is
merged it will still be necessary to write a few docs, including:
  - Revise `farmdata2_modules/README.md`
  - Write `farmdata2_modules/fd2_example/README.md`
  - Write `farmdata2_api/README.md`

New issues should be created for each of these when this PR is merged.

Closes #609 
Closes #610
Closes #572 

Changes to the INSTALL.md document were 
Co-authored-by: John MacCormick <jmac@dickinson.edu>
as adapted from his PR #612.

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.