Backend and Frontend Fixes for Transplanting Report (#615)
__Pull Request Description__

Closes #562. This issue makes some changes to the `updateRow` function
located in the Transplanting Report HTML file
[[here](https://github.com/FarmData-2-Dev-Team-2022/FarmData2/blob/fixTransplantingBackend/farmdata2/farmdata2_modules/fd2_barn_kit/transplantingReport/transplantingReport.html)].
These changes are not accompanied by any Cypress e2e testing as that is
a separate issue to be addressed in issue #578. As a result, these
changes have been tested by sight using Hoppscotch.

Additionally, this PR makes a quick fix of the filters for both the
Seeding Report and Transplanting Report where editing once would
permanently disable the filters until the page was refreshed because the
name of the edit cancellation event was changed and the pages were not
updated accordingly to reflect this change.

This PR also brings up the topic of #585 (and #466) of what should be
edit-able. I disabled the ability to edit `bed feet` in the
Transplanting Report because transplanting logs don't have quantity data
for `bed feet` but also enabled the ability to edit `row/bed` because
that does exist in the logs' quantity data. However, should users be
allowed to edit that data? This is beyond the scope of this PR but
wanted to quickly mention this quick thought that occurred to me whilst
making this fix.

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.