Fixed broken FD2School links (#616)
__Pull Request Description__

The activities in the `FD2School/materials` directory were renumbered
but the links were not updated. This PR updates the links to match the
filenames of the materials.

Co-authored-by: John MacCormick <jmac@dickinson.edu>

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.