Clarified and fixed CustomTable data-cy attributes for input elements (#623)
__Pull Request Description__

The CustomTableComponent `data-cy` attributes for input elements did not
uniquely identify the input element. In particular, if two elements had
the same type of input they would have had the same `data-cy` attribute
value. They now have the value `ri-columnHeader-input` where the
`columnHeader` is the header of their column in the table. This matches
what is done for cells when the table is not in edit mode.

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.