Added the pageLoaded pattern to barnKit and fieldKit (#624)
__Pull Request Description__

The pageLoaded pattern has been applied to the barnKit and fieldKit
pages. This pattern places a hidden element in the page that has its
value set to true when all of the API calls in the `created` method have
completed. This is used in testing to ensure that the page has fully
loaded before the tests are run.

This is important because when pages are loading information in the
`created` method, fast running tests were completing and the next test
was starting before those API requests had completed or even been
initiated. This was generating flakiness in the test suite. It would run
fine one time and not the next. Waiting for the page to load introduces
delays, but removes the flakiness.

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.