Encapsulating Button Cases for the CustomTableComponent (#626)
__Pull Request Description__

Closes #599. This PR adds a computed property to the
`CustomTableComponent.js` to remove the multiple locations in the
template checking when to make the selection bar appear in the table.
Because it would be easy to miss these locations when adding new button
cases, this computed method serves as a general locations for adding
this button locations without worrying about making changes to the
template.

Because this doesn't alter functionality, no additional testing was
added. However, the table's unit test was updated because the Cypress
tests for the CSV download broke with the reorganization of the project.
The path is currently fixed. This is fine because the structure of the
project doesn't change frequently but should we consider putting more
resources into making it dynamic to not worry about this in the future?
This is outside the scope of the issue this PR aims to resolve, but I
could make an issue if it's worthwhile.

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.