Surpressing Standard Name Error (#627)
__Pull Request Description__

Closes #565. Adds a simple simple `@` in front of the cmd command that
maps the user ID's and names as a preprocessing function. This error
would appear when users attempted to access a page without being logged
into the page. The warning no longer appears due to this surpression and
a comment explaining the `@` placement is now present in the module
files.

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.