Fix typos in documentation (#638)
__Pull Request Description__

This PR fixes minor typographical errors, such as misspelled words,
omitted or extra words, and missing punctuation, in all of the Markdown
doc files.

These changes will improve the clarity of the documentation for
FarmData2 users.

__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.