Generalization of the ErrorBannerComponent  (#622)
__Pull Request Description__
This PR closes #601. This pull request generalizes the functionality of
the `ErrorBannerComponent` into the newer `BannerComponent` which can
display error, success, and generic message banners. Additionally, the
refactored component can support timeout (instead of clicking an _X_ to
close the component, it'll collapse itself after 5secs) and is has a
sticky style which makes it appear on screen without the intrusive auto
scroll.

This PR will not remove the pre-existing `ErrorBannerComponent` from the
project. That is something that will need to be done in a separate PR
that also replaces those instances with the new `BannerComponent`.

This PR is **NOT** ready to be merged into main. 
Prior to merging, the following items must be completed. 
- [x] Complete code review 
- [x] Remove all instances of the `ErrorBannerComponent` from the
ui.html page.
- [x] Update the end-2-end test for the ui.html page. 
- [x] Remove all lingering debugging code from the PR. This includes
unnecessary comments.
- [x] ~~Regenerate documentation~~ (This is generated locally now)


Edit: Closes #635. 

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.