Db test waits (#646)
__Pull Request Description__

Modified the `dbtest` examples so that they use the `cy.waitForPage()`
pattern that will be typical of real use cases.

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.

---------

Signed-off-by: Foogabob <80643562+Foogabob@users.noreply.github.com>
Co-authored-by: pranavm2109 <98339495+pranavm2109@users.noreply.github.com>
Co-authored-by: Alexandrialexie <brownale@dickinson.edu>
Co-authored-by: andrewscheiner53 <scheinea@dickinson.edu>
Co-authored-by: sidlamsal <98344853+sidlamsal@users.noreply.github.com>
Co-authored-by: EliasBerhe <eliastes123@gmail.com>
Co-authored-by: Foogabob <locker@dickinson.edu>
Co-authored-by: Foogabob <80643562+Foogabob@users.noreply.github.com>
Co-authored-by: Michael K <56452607+Mikek16@users.noreply.github.com>
Co-authored-by: nathang15 <nguyenat@dickinson.edu>
Co-authored-by: infantlikesprogramming <hadu@dickinson.edu>