Created Swagger Doc for fd2_api; Refactored of cropmap and areamap in seeding report; Updated FarmData2API.js (#650)
__Pull Request Description__

See PR #630 for a full description of changes.

Closes #630, #597, #606, #607

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.

---------

Co-authored-by: JingyuMarcelLee <leemar@dickinson.edu>
Co-authored-by: Jingyu (Marcel) Lee <46755208+JingyuMarcelLee@users.noreply.github.com>