Fixed Total Bedft not Rounded in Seeding Report (#678)
__Pull Request Description__

Closes #657. This PR makes small changes to two computed methods in the
Seeding Report page that concerns how the total bed feet is returned for
the page to display in the direct seeding summary. Since logs can make
the number a long decimal numbers, the return in these methods now
rounds the value to the nearest integer to resolve this issue on the
frontend.

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.