Migrates: Main fd2 tabs test #228 (#661)
…l/228
This pull request was orignially created on
[FD2School-FarmData2](https://github.com/DickinsonCollege/FD2School-FarmData2).
Link to the original pull request:
[#288](https://github.com/DickinsonCollege/FD2School-FarmData2/pull/228)

Authors of the original Commit are listed here as co-Authors:
Co-authored-by: jamesng5 <nguyenja@dickinson.edu>
Co-authored-by: vuphuongha <havu0910@gmail.com>
Co-authored-by: tainguyen103 <tainguyen103200@gmail.com>

Partially Addresses #662

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.

Co-authored-by: James Ng <98336522+jamesng5@users.noreply.github.com>