Migrates: Test Barn Kit Seeding Report columns by seeding type #210 (#663)
Co-authored-by: nathang15 <nguyenat@dickinson.edu>
Co-authored-by: infantlikesprogramming <hadu@dickinson.edu>

__Pull Request Description__

This pull request was orignially created on
[FD2School-FarmData2](https://github.com/DickinsonCollege/FD2School-FarmData2).
Link to the original pull request:
[210](https://github.com/DickinsonCollege/FD2School-FarmData2/pull/210)

Partially Addresses
https://github.com/DickinsonCollege/FarmData2/issues/662

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.

Co-authored-by: Michael K <56452607+Mikek16@users.noreply.github.com>