Migrates: Seeding Input: Tests for Submit Button Behavior (#208) (#665)
Co-authored-by: Foogabob <locker@dickinson.edu>
Co-authored-by: EliasBerhe <eliastes123@gmail.com>

__Pull Request Description__

This pull request was orignially created on
[FD2School-FarmData2](https://github.com/DickinsonCollege/FD2School-FarmData2).
Link to the original pull request:
https://github.com/DickinsonCollege/FD2School-FarmData2/pull/208

Partially Addresses
https://github.com/DickinsonCollege/FarmData2/issues/662

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.

Signed-off-by: Foogabob <80643562+Foogabob@users.noreply.github.com>
Co-authored-by: sidlamsal <98344853+sidlamsal@users.noreply.github.com>
Co-authored-by: Foogabob <locker@dickinson.edu>
Co-authored-by: EliasBerhe <eliastes123@gmail.com>
Co-authored-by: Foogabob <80643562+Foogabob@users.noreply.github.com>