Migrates: Tested the Seeding Report Crop Filter (#196) (#666)
Co-authored-by: andrewscheiner53 <scheinea@dickinson.edu>
Co-authored-by: pranavm2109 <mishrap@dickinson.edu>

__Pull Request Description__

This pull request was orignially created on
[FD2School-FarmData2](https://github.com/DickinsonCollege/FD2School-FarmData2).
Link to the original pull request:
https://github.com/DickinsonCollege/FD2School-FarmData2/pull/196

Partially Addresses
https://github.com/DickinsonCollege/FarmData2/issues/662

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.

Co-authored-by: Alexandrialexie <98338885+Alexandrialexie@users.noreply.github.com>
Co-authored-by: andrewscheiner53 <scheinea@dickinson.edu>
Co-authored-by: pranavm2109 <mishrap@dickinson.edu>