Migrates: Testing Transplanting Report Defaults (#197) (#672)
* Added spec.js file for testing transplating report

* Tested page header and section label

* Tested if the generate button has correct value and is enabled

* The report table is not visible

* Tested default start, end dates using dayjs

* Changed data-cy atrribute name for - Set Dates

* click() is removed when getting the default dates

* Made final changes for the test

* Made final changes to the test

__Pull Request Description__

This pull request was orignially created on
[FD2School-FarmData2](https://github.com/DickinsonCollege/FD2School-FarmData2).
Link to the original pull request:
https://github.com/DickinsonCollege/FD2School-FarmData2/pull/197

Partially Addresses
https://github.com/DickinsonCollege/FarmData2/issues/662

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.

Co-authored-by: Udval Enkhtaivan <74579865+udvale@users.noreply.github.com>