Migrates: 227 testing the default values of the seeding report (#677)
* updated fd2 branch + tests

* Updated tests

* test for Barn-Kit seeding report default

* updated tests

* updated tests

* "Made minor corrections to tests"
>
>
Co-authored-by: Sai Atluri <atluris@dickinson.edu>

* Updated date selection header test

* Updated date selection header test

* Corrected minor minor merging error

* Fixed formatting

Co-authored-by: Melantha-Chen <yuxuan.chen.1031@outlook.com>

__Pull Request Description__

This pull request was originally created on
[FD2School-FarmData2](https://github.com/DickinsonCollege/FD2School-FarmData2).
Link to the original pull request:
https://github.com/DickinsonCollege/FD2School-FarmData2/pull/227

Partially Addresses
https://github.com/DickinsonCollege/FarmData2/issues/662

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.

---------

Co-authored-by: saiatl2354 <75409740+saiatl2354@users.noreply.github.com>
Co-authored-by: Melantha-Chen <yuxuan.chen.1031@outlook.com>