Change default seeding type according to existing logs (#680)
__Pull Request Description__
This PR changes the `selected-val` of the Seeding Type Dropdown from
'All' to the `getDefaultSeedingSelection` function that returns the
appropriate seeding type for the available logs.

Closes #658 

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.