Refactored dev/startup.bash Docker container (#682)
__Pull Request Description__

I mostly reworded some of the comments to be clearer
and used a single 'sudo' command with a here-document to reduce
footprint. It's a small portion, I would imagine it will take a while to
actually speed up the container but every little bit helps!

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.

Signed-off-by: Ty Chermsirivatana <chermsit@dickinson.edu>