Implementation of Selective Containers at Launch (#684)
__Pull Request Description__

This PR should resolve #575 ~~and is currently in a draft phase~~.
Basically this is going to add a phpmyadmin profile, then the PR
modifies `fd2-up.bash` to include flags such as `--no-dev` and
`--no-phpmyadmin.` ~~Will test and work before recommending a review.~~

__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.

---------

Signed-off-by: Ty Chermsirivatana <chermsit@dickinson.edu>