Add  cypress logging functionality (#692)
__Pull Request Description__

This add small feature to help log information such as request and
response to terminal when using cypress test in headless mode . To log
information, use command cy.task( "log", <data>) to print to terminal.

Closes #686

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.