Split FarmOSAPI test file (#688)
__Pull Request Description__

Create a test folder and divide tests in `FarmOSAPI.js` into smaller
chunks

closes #651 
---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.