revert changes to fd2-down.bash to work on MacOS too (#694)
__Pull Request Description__

The `COMPOSE_PROFILE` environment variable was not working on MacOS. So
reverted the `fd2-down.bash` script to use the `stop` and `delete`
commands instead.

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.