add Bug Reporting and Feature Requests sections and links to those (#698)
__Pull Request Description__

Closes #94

Added sections like Bug Reporting and Feature Requests in
Contributing.md and added approporiate links to waysToContribute.md file
for bug reporting and feature request sections

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.